web/requirements.txt
alembic==0.9.7
bcrypt==3.1.4
certifi==2018.1.18
cffi==1.11.4
chardet==3.0.4
click==6.7
Flask==0.12.2
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.3
Flask-Login==0.4.1
Flask-Migrate==2.1.1
Flask-SQLAlchemy==2.3.2
gunicorn==19.7.1
httpie==0.9.9
idna==2.6
itsdangerous==0.24
Jinja2==2.10
Mako==1.0.7
MarkupSafe==1.0
psycopg2==2.7.3.2
pycparser==2.18
Pygments==2.2.0
PyJWT==1.5.3
python-dateutil==2.6.1
python-editor==1.0.3
requests==2.18.4
six==1.11.0
SQLAlchemy==1.2.1
urllib3==1.22
Werkzeug==0.14.1
